shinytheme {shinythemes}	R Documentation

Return the URL for a Shiny theme

Description

The result of this function should be used as the theme argument for bootstrapPage, fluidPage, navbarPage, or fixedPage.

Usage

shinytheme(theme = NULL)

Arguments

theme
Name of a theme.

See Also

The main shinythemes page for information about available themes and more detailed examples.

Examples

## Not run: 
##D shinyApp(
##D   ui = fluidPage(theme = shinytheme("united"),
##D     ...
##D   ),
##D   server = function(input, output) { }
##D )
## End(Not run)

[Package shinythemes version 1.2.0 Index]